Drop images on the floor (don't retry) when they fail to save.
Retrying was causing both memory bloat (to store the image until it saves) and CPU lag (trying to process the same images over and over. In the future we should try to add UIImages to a queue and process them on a dedicated background thread (instead of the main thread).